feat: Add possibility for IoU margin between pos and neg anchor boxes
Added the possibility to make certain anchor boxes neither positive nor
negative by specifying a maximum IoU threshold that, if exceeded,
disqualifies the box as a negative.

It is useful to exclude such boxes from the training loss whose IoU is
not quite high enough to qualify as a positive match, but almost high
enough, creating a margin between clear negatives and clear positives.