fix: Catch deformed ground truth boxes to avoid division by zero
Catch bad ground truth boxes that have a width or height of zero and ignore them. This is important because such boxes would break the log computation in `encode_y` (and besides, of course, boxes with zero width or height don't make any sense and were likely just labeling mistakes when the data was created).