feat: Add confidence threshold option to `decode_y()`
Added and option to `decode_y` that allows to set a minimum classification confidence that a given box needs to exceed in order to count as a positive prediction. Before, any box whose maximal confidence was in a non-background class counted as a positive prediction.

The output of the function was also extended: The function now outputs the classification confidence in addition to the former output.