feat: Allow last box scaling factor to be set explicitly
Before, the list of box scaling factors contained one element per classifier layer. If `two_boxes_for_ar1` is `True`, the size of the second aspect-ratio-1 box for a given layer is computed using geometric mean of the scaling factor for that layer and the scaling factor of the successive layer. Up until now, for the very last classifier layer, because there is no successive layer, this second scaling factor was hard-coded to 1.0. Now it can be set explicitly, which is why the list of scaling factors now needs to have one additional element, that is, there is now always one more scaling factor than there are classifier layers.

The training notebooks have been updated accordingly.