feat: Add random crop feature
Added a feature to generate random crops. The target crop size can be arbitrary, i.e. the crop size can also be larger than the original image, not only smaller. This allows to generate images of arbitrary size and aspect ratio without having to distort the original aspect ratio of the image.

This feature is a great tool for data augmentation, but more importantly it was needed to deal with datasets in which the images vary in size. Taking random crops of a fixed size is the proper solution to this issue.