feat: Add XML parser for Pascal VOC datasets
The generator was slightly refactored to accommodate the possibility for multiple parsers to parse different types of annotation formats. Before, the constructor automatically parsed the labels from a suitable CSV file.

Now, the constructor only sets a few variables common to all parsers. The image lists and label lists are then parsed using one of the parser methods.