fix: Fix broken batch item removal after failed random cropping
If random cropping with `min_1_object`  is enabled, a batch item will be removed from the batch if no random crop can be found that contains at least one object after the maximum number of trials.

Until now, there was a bug in the code that led to the wrong batch items being removed from the batch if more than one batch item was to be removed. The indexing for the subsequent removal of several batch items has now been corrected so that the right batch items will be removed.