feat: Add coordinate normalization and original model output processing
Added an option to normalize box coordinates to be in [0, 1], i.e. relative to the image height and width. Note that whether you normalize or not does not make a difference for training, since all targets are being divided by the anchor box widths or heights, so the normalization cancels out.

Also added the original model prediction decoding procedure from the SSD paper in `decode_y()`, which applies a very low confindence threshold first (0.01), performs non-maximum suppression per class second, and keeps the top-k remaining highest confindence boxes third.

The procedure that does not do NMS per class and is a lot more efficient in turn has been moved to `decode_y2()`.