fix: Set correct input for anchor box layers
The inputs to the anchor box layers were incorrectly set to conv layers of the base network, when it should instead have been set to localization predictor layers. This has now been fixed.

This mistake had no effect, because the padding of the localization predictor layers is set to "same", and so the width and height of each localization layer is the same as that of the respective underlying conv layer that feeds into it, so in this particular case it doesn't matter from which of the two alternatives the anchor box layer gets the tensor shape information, but of course this is still an important fix to maintain the logical consistency and generality of the setup.