feat: Support processing of multiple dataset paths in XML parser
Added support so that the XML parser can process multiple dataset paths at once instead of only one path. This is useful in order to train on multiple datasets that are located in different directories instead of only being able to specify one directory. For example, when training on Pascal VOC, you might want to train the model on VOC2007, VOC2007_Test, and VOC2012 all at the same time.

This change required a tiny bit of refactoring and API changes, but it should be straight forward using the documentation.

Essentially, all paramters that accepted strings (representing paths) before now accept lists of strings instead.