doc: Fix documentation regarding model output shape
The documentation regarding the model output shape and format was incorrect in multiple classes and functions. This has been fixed now. The olde, erroneous documentation said that the model outputs a tensor of shape `(batch_size, #boxes, #classes + 8)`, when it actually outputs a tensor of shape `(batch_size, #boxes, #classes + 12)`, where the last four entries in the last axis are the four variance values.