feat: Add support for loading complete models instead of weights only
Finally added support to load complete models. You can now load saved models using the `load_model()` function rather than having to build the model and then load weights into it. The advantage of loading the complete model is that the current training state is saved along with the model.