feat: Support generating batches of images only, no labels
Previously the batch generator was only able to generate batches of both images and labels, which made it unsuitable to be used for inference on a test dataset for which no ground truth is available. This has now been fixed: You now have the option to load just images and no labels, in which case it is not necessary to call any of the parser methods. Simply pass a suitable image list in the `BatchGenerator` constructor (see documentation).

Also added the possibility to turn off shuffling, which can be useful for debugging or prediction.