fix: Make loss independent of batch size
Up until now, the training loss was dependent on the batch size. To be precise, it was inversely proportional to it. It wasn't purposely designed like this, it was just a consequence of the fact that Keras divides the loss by the batch size regardless of the loss function a model uses, but the SSDLoss function already averages the loss internally (not by the batch size, but by the number of positive ground truth boxes in the batch, a number which is of course highly correlated to the batch_size). This resulted in the loss practically being divided by the batch size (or a number proportianl to it) twice, leading to the loss being dependent on the batch size. I thought this wasn't problematic, so I didn't worry about it, but I have now realized that it can lead to issues in certain cases.

The problem occurs to varying degrees when the number of training images is not an integer multiple of the chosen batch size (which is almost always the case of course). In this case, the very last batch of an epoch will contain fewer images than the batch size. Since the loss was inversely proportional to the actual size of the batch, this caused a spike in the loss each time at the point where the smaller batch occurred.

This has now been fixed. SSDLoss now multiplies the loss by the batch size, effectively reverting Keras' subsequent division by it.