feat: Add diagnostics option for `encode_y()`
Included a diagnostics option by which `encode_y()` can output a tensor with matched anchor boxes, i.e. where only the classes are encoded, but the coordinates are the anchor box coordinates instead of the ground truth coordinates. This can be very useful for visualizing which anchor boxes are being matched to which ground truth boxes.