feat: Switch to He normal initializer in all conv layers
Changed all conv layer initializers to He normal initialization instead of the default Glorot uniform initialization.

This switch was made based on the results of training an untrained SSD300 for 15000 steps with each initializer. The training loss of the model initialized with He normal decreased significantly faster than that of the model initialized with Glorot uniform in the first few thousand training steps, after which the training losses of the two models converged towards each other, but the loss of the He-initialized model was still consistently lower. This indicates that He initialization makes the model converge faster, although the difference might become smaller the longer the model is being trained.