feat: Make anchor computation more flexible, replace input norm layer
Up until now, the anchor box center points were computed by dividing the image into identical rectangular cells and making the center points of those cells the center points of the anchors. This approach makes sense, but it leaves no flexibility to tweak the resulting grid. In particular, it made it impossible to reproduce the exact anchor box distribution of the original SSD implementation for all layers (it was exact for some layers, but not for all). Because of this, the computation of the center points has now been changed to incorporate the two parameters of the original SSD implementation, namely "step" and "offset". This new computation formula follows the original implementation exactly and at the same time retains the flexibility to go back to the previous computation formula of this Keras implementation by simply leaving both "steps" and "offsets" `None`.

I'm in the process of porting the original trained SSD weights and this change was necessary in order to reproduce the results of the original implementation. The computation before wasn't wrong in any way, but in order to use the original trained weights, the anchor boxes in this Keras implementation must be computed precisely the same way as in the original Caffe implementation.