feat: Support "corners" coordinate format, change anchor computation
Added support for the `(xmin, ymin, xmax, ymax)` ("corners") format that the original Caffe implementation uses.

Changed the computation of the anchor box center points to follow the original Caffe implementation exactly. The previous computation formula of this Keras implementation wasn't wrong in any way, but I'm in the process of porting the original Caffe weights and in order to reproduce the original results, the anchor box computation needs to be identical to the Caffe implementation.