feat: Add 'swap_channels' option since SSD300 requires BGR input
Added an option to both the batch generator and the SSD300 build function that reverses the color channel order of the input images.

The original Caffe implementation of SSD assumes BGR input.

Of course the idea is to use the `swap_channels` option either in the SSD300 model build function or in the batch generator, not in both. It just seemed sensible to add the option at both levels.