fix: Adjust padding for use with original trained weights
Nothing was wrong with the padding before per se, the model worked absolutely fine if you trained it from scratch.

However, in order to use the model with the ported trained weights from the original Caffe implementation, I needed to adjust the padding a bit so that every pixel is in the exact same place as in the original model.

This turns out to be incredibly important. If the padding on either side of the output of a given layer is off even by one pixel, the predicted boxes for all affected predictor layers will be awfully far off.

With the padding now fixed, the model's predictions are really good when using the original trained weights.