refactor: Make box encoding more efficient
The computation of the anchor box grid for each predictor layer has been moved into the constructor. These anchor box grids only have to be computed once since they depend only on the model configuration, not on the input data, so it was a waste to re-compute them for every new batch that needs to be encoded (the way it was before).