chore: Switch the preset decoder to the original Caffe version
This implementation provides two different functions to decode raw model output, one follows the decoding procedure of the original Caffe implementation, the other is my own version. Up until now, my version was preset in this notebook, but now I've switched it to the original version.

Also made a few minor style and documentation changes, most noteworthily explained the difference in how the two decoders perform NMS and what that means in practice.