feat: Define `n_classes` to be the number of positive classes
Previously, any function that expects an `n_classes` argument expected the passed value to include the background class, e.g. you had to pass 21 for Pascal VOC or 81 for MS COCO.

This was unintuitive and a multiple users tripped over this subtlety. There was no strong reason why it should be this way, so now it's been changed. `n_classes` is now defined to be the number of positive classes, i.e. excluding the background class, e.g. 20 for Pascal VOC or 80 for MS COCO.