refactor: Simplify generator output selection
The `generate()` now has an argument `returns`, which is a set of strings. In order to specify that the generator shall include a specific item in the yielded outputs, just include the respective keyword string in the `returns` set. This way, `returns` let's you specify what outputs you need from the generator. 

This is simpler and more flexible than the previous state, where only some of the generator's outputs were user-controllable via the `diagnostics` option. It was a pretty messy system in which it was unclear what the generator returns under which circumstances and why. The `diagnostics` option has been removed in turn since it is made obsolete by the new `returns` argument.