chore: Enable mean normalization and channel swapping by default
The original SSD300 and SSD512 models expect their input to be BGR images, not RGB, and to be per-channel mean-normalized by the values [104, 117, 123] (for both Pascal VOC and MS COCO), so the default parameters in the model builder functions have now been set accordingly.

mAP evaluation on MS COCO also confirmed that the model performance using the ported original weights is significantly higher with input mean-normalized and in BGR channel order.