fix: Make sure image arrays always have ndims==3
This fix concerns an edge case.

If a dataset consists of grayscale (single-channel) images and `convert_to_3_channels` is set to `False`, the generated batch of images will be a 3-dimensional array, but Keras requires a 4-dimensional array as input.

This has now been fixed: All images will now be expanded to have `ndmins == 3` if that is not already the case.