chore: Update dataset with train/val split, include history graph
Updated the processed Udacity road traffic dataset that is linked in the notebook with a train/val split. Before, there was just one single labels file that contained all ground truth labels.

Also included a graph of the training history to give other users a reference of what the training should look like over the first 30 epochs with parameters as set in the notebook.