feat: Always include background class when random sampling
Previously, in the case of random sub- or upsampling, all indices were selected at random, which means you couldn't be sure that the background class was included in the sampled tensor. Since we can assume that you will always want the background class, this has now been changed: All indices except the background class are sampled randomly, but the background class will always be included.