feat: Allow model to be created with or without DecodeDetections layer
The constructor now offers the possibility to create the model with or without the `DecodeDetections` layer. In order to train the model, it must be created in `training` mode, which omits the `DecodeDetections` layer. In order to use a trained model for inference, however, it can be created in `inference` mode, in which case it outputs decoded and filtered predictions instead of raw predictions. Using the `DecodeDetections` layer at inference time significantly speeds up prediction.