fix: Change `iou_threshold` argument from tensor to float
Apparently (and annoyingly), the parameter `iou_threshold` in the function `tf.image.non_max_suppresion()` specifically requires a float as input in some versions of TensorFlow. In other versions it can also handle a tensor of type float. In order to be compatible with all versions, the input to this parameter has now been changed from a 1-element float tensor to just a float.