feat: Set channel order explicitly in `swap_channels`
Previously, `swap_channels` could only be set to `True` to reverse the input image channel order or to `False` to do nothing. Now the channel order is set explicitly via a list.

Also, renamed the `limit_boxes` argument to `clip_boxes` in accordance with the same name change in other modules.