chore: Remove internal coordinate format options
The difference between the internal coordinate formats that the model can learn and the coordinate formats of the input and output data has recently caused some confustions and led to a bunch of questions, so that the option to change the internal coordinate format is now entirely omitted in the tutorial notebooks. It's not relevant at all to change the internal coordinate format and it only led to confusion, so it's probably best not confront users with this option to begin with.