fix: Ensure `pred` is an array even if empty
This should only have caused problems in rare edge cases, but the list of decoded predictions for a given batch item, `pred`, should be a Numpy array even if there are no predictions at all. Previously, if `pred` was empty, it wasn't converted from a list to an array. Fixed now.