feat: Upgrade IoU function
Upgraded the IoU function with a new `mode` argument. In 'element-wise' mode, the function works exactly the same as before. In 'outer_product' mode, returns a matrix with the IoUs of all possible combinations between the first and the second set of boxes.

Also refactored the code: The intersection area computation part is now its own function, because computing intersection areas is generally useful functionality by itself.