fix: Make XML robust to varying tag order
It turns out that the XML parser had a bug: For each ground truth object that it parsed, it took the first occurring bounding box for that object. This behavior led to the correct bounding boxes being parsed for the vast majority of the ground truth bounding boxes, but in some cases, particularly for the Pascal VOC object class "Person", there is more than one bounding box per object (body parts) and what's worse, the main bounding box is not always the first bounding box among the tags.

This has now been fixed. The parser always gets the correct bounding box for each object.