fix: Treat bounding box border pixels correctly
It is strongly recommended to pull this fix if you're on one of the recent commits.

A recent commit made a subtle adjustment to how border pixels of bounding boxes are treated by default. This change had an unexpected negative impact on the training performance. This has now been fixed and the training performance is back to normal.